add support for /ResolveMatches to client
it appears a service can chose to send resolution data in its /ProbeMatches
reply OR not. when not, the client would need to explicitly request
/Resolve (for which we already have support) and get the resolution data
in the /ResolveMatches reply.

IOW: if probeMatchReceived() signals with a service that has no xaddrs
it's not yet resolved and needs running through sendResolve() and
get the service extended through the reply before being actually useful
outside probe discovery.

e.g. https://github.com/christgau/wsdd behaves this way